VHDL transmitter and receiver simulation
Serial data transmitter is a circuit which takes in parallel data and transmits out serial data which
certain constrains. It has a seven bits data input, din (6:0), and a control signal start_transmit. A
clock signal clk is also needed at the input. It has one output, serial_out, which is a single bit used to
send out the serial data.
- When no bits are sent, the serial_out is ‘0’. The sent serial data consists of ten bits. The first bit is a
start bit, which must be high to inform the receiver with presence of a new serial data. The next
seven are the actual data bits, din, where the least significant bit of din is send first. The ninth bit is a
parity bit, whose status must be ‘0’ if the number of ones in data is even or ‘1’ otherwise. Finally, the
tenth is a stop bit, which must be sent high.
- When the start_transmit signal is asserted ‘1’, the circuit reads the input data, din, calculate the
parity bit and start sending the ten bits serially. A new bit is send at the rising edge of the clk. When
finishing, the circuit must return the serial_out to ‘0’ again